Warehouse crud

+++ src/app/dashboard/humanResources/WorkersTable.tsx
@@ -14,11 +14,12 @@ import { Toast } from "@/helpers/customAlert";
 import { startAddWorker, startDeleteWorker, startUpdateWorker, workersStartLoading } from "@/actions/workers";
 import { CreateWorkerForm } from "./CreateWorkerForm";
 import { EditWorkerForm } from "./EditWorkerForm";
+import Swal from "sweetalert2";
 
 // TODO: Hacer que al editar un trabajador y volver a presionar editar en el formulario aparezcan los datos actualizados
 
 interface DataType {
-  _id: string,
+  _id: string;
   key: string;
   name: string;
   CI: number;
@@ -59,7 +60,6 @@ const WorkersTable: React.FC = () => {
         title: "Seleccione un trabajador a editar",
       });
     }
-    
   };
 
   const onCreate = (values: any): void => {
@@ -68,7 +68,7 @@ const WorkersTable: React.FC = () => {
   };
 
   const onEdit = (values: any): void => {
-    dispatch(startUpdateWorker(selectedRow?._id!,values.name, values.CI, values.address, values.role, values.phoneNumber, values.bankAccount));
+    dispatch(startUpdateWorker(selectedRow?._id!, values.name, values.CI, values.address, values.role, values.phoneNumber, values.bankAccount));
     setSelectedRow(undefined);
     setEditModal(false);
   };
@@ -81,7 +81,20 @@ const WorkersTable: React.FC = () => {
 
   const handleDelete = () => {
     if (selectedRow) {
-      dispatch(startDeleteWorker(selectedRow?.CI));
+      Swal.fire({
+        title: "Eliminar Trabajador",
+        text: "El trabajador seleccionado se borrará de forma permanente",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        cancelButtonText: "Cancelar",
+        confirmButtonText: "Eliminar",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          dispatch(startDeleteWorker(selectedRow?.CI));
+        }
+      });
     } else {
       Toast.fire({
         icon: "error",
@@ -256,24 +269,25 @@ const WorkersTable: React.FC = () => {
 
   return (
     <>
-      <div className="flex h-14 w-full bg-white rounded-md shadow-md mb-4 items-center pl-4 gap-2">
-        <div
+      <div className="flex h-14 w-full bg-white-100 rounded-md shadow-md mb-4 items-center pl-4 gap-4">
+        <button
           onClick={handleNew}
-          className="bg-success-500 w-[6rem] h-[2.5rem] flex items-center p-1 font-black text-white cursor-pointer justify-center gap-2 rounded-md hover:bg-success-600 ease-in-out duration-300"
+          className="bg-success-500 w-[6rem] h-[2.5rem] flex items-center p-1 font-black text-white-100 cursor-pointer justify-center gap-2 rounded-md hover:bg-success-600 ease-in-out duration-300"
         >
           <PlusOutlined />
           Nuevo
+        </button>
+        <div className="flex gap-2">
+          <button className="cursor-pointer" id="edit_worker_btn" onClick={handleEdit}>
+            <EditOutlined className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-white-600 ease-in-out duration-300" />
+          </button>
+          <button className="cursor-pointer" id="delete_worker_btn" onClick={handleDelete}>
+            <DeleteOutlined className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-white-600 ease-in-out duration-300" />
+          </button>
+          <button className="cursor-pointer" onClick={() => dispatch(workersStartLoading())}>
+            <ReloadOutlined className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-white-600 ease-in-out duration-300" />
+          </button>
         </div>
-        <button className="cursor-pointer" id="edit_worker_btn" onClick={handleEdit}>
-          <EditOutlined className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-background_light ease-in-out duration-300" />
-        </button>
-        <button className="cursor-pointer" id="delete_worker_btn" onClick={handleDelete}>
-          <DeleteOutlined className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-background_light ease-in-out duration-300" />
-        </button>
-        <ReloadOutlined
-          onClick={() => dispatch(workersStartLoading())}
-          className="w-[2rem] h-[2rem] text-xl rounded-full hover:bg-background_light ease-in-out duration-300"
-        />
       </div>
 
       <CreateWorkerForm open={createNewModal} onCancel={() => setCreateNewModal(false)} onCreate={onCreate} />
@@ -283,7 +297,7 @@ const WorkersTable: React.FC = () => {
         size="middle"
         columns={columns}
         dataSource={data}
-        pagination={{ position: ["bottomCenter"], pageSize: 15 }}
+        pagination={{ position: ["bottomCenter"], pageSize: 10 }}
         rowSelection={{
           type: "radio",
           ...rowSelection,
